requirements.txt
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
cloudpickle==3.0.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
deap==1.4.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
fonttools==4.43.1
fsspec==2023.10.0
gym==0.26.2
gym-notices==0.0.8
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickle5==0.0.11
pickles==0.1.1
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.9.7
Pygments==2.16.1
pygraphviz==1.11
pyparsing==3.1.1
pyproject-toml==0.0.10
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
rpds-py==0.12.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
toml==0.10.2
torch==2.1.0
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.9
zipp==3.17.0
